+++ CHANGELOG.md
@@ -13,6 +13,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - The `@cumulus/ingest/util.lookupMimeType` function now returns `undefined`
   rather than `null` if the mime type could not be found.
 - The `@cumulus/ingest/lock.removeLock` function now returns `undefined`
+- The `@cumulus/ingest/granule.generateMoveFileParams` function now returns
+  `source = undefined` and `target = undefined` if either could not be
+  determined. Previously, `null` had been returned.
 
 ### Added
 
@@ -23,6 +26,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - Fixed bug with request headers in `@cumulus/launchpad-auth` causing Launchpad token requests to fail
 
+### Removed
+
+- The `@cumulus/ingest/granule.copyGranuleFile` function has been removed
+- The `@cumulus/ingest/granule.moveGranuleFile` function has been removed
 
 ## [v2.0.0] 2020-07-23